requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.1
Django==1.11.3
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-filter==1.0.4
django-imagekit==4.0.1
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
Markdown==2.6.8
olefile==0.44
pilkit==2.0
Pillow==4.2.1
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
whitenoise==3.2
